refactor(xml): extract XmlCaptureSession, XmlExtractionState, XmlOutputEmitter from XmlFilterCommand
CaptureSession と ExtractionState を package-private クラスとして分離することで TooManyMethods PMD 違反を解消し、@SuppressWarnings("PMD.TooManyMethods") を不要にした。

- XmlCaptureSession: XMLEventWriter のライフサイクル管理（start/add/finish/abort）
- XmlExtractionState: 走査中の可変状態（パス・深度・出力状態）と XmlCaptureSession の所有
- XmlOutputEmitter: 出力フォーマット（単一要素 vs ラップ出力）の静的ユーティリティ
